requirements.txt
amqp==5.0.2
asgiref==3.3.1
billiard==3.6.3.0
boto3==1.16.43
botocore==1.19.43
celery==5.0.5
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
Django==3.1.4
django-filter==2.4.0
djangorestframework==3.12.2
flower==0.9.7
humanize==3.2.0
jmespath==0.10.0
kombu==5.0.2
Markdown==3.3.3
prometheus-client==0.8.0
prompt-toolkit==3.0.8
pycurl==7.43.0.5
python-dateutil==2.8.1
pytz==2020.5
s3transfer==0.3.3
six==1.15.0
sqlparse==0.4.1
tornado==6.1
urllib3==1.26.2
vine==5.0.0
wcwidth==0.2.5
